delete images with deleting products

+++ routes/products.js
@@ -5,7 +5,6 @@ const User = require('../models/userModel');
 const Product = require('../models/productModel');
 const Image  = require('../models/imageModel')
 const ImageChunk  = require('../models/imageChunkModel')
-console.log('asdfsd')
 // get all products (from 0 to 15 with a skip)
 router.get('/',async(req,res)=>{
     try {
@@ -87,8 +86,11 @@ router.delete('/:_id',authenticate,adminAuthenticate,async(req,res)=>{
     try {
         let {_id} = req.params;
         let product = await Product.findOne({_id});
-        let image = await Image.findOneAndDelete({filename:product.image});
-        await ImageChunk.deleteMany({files_id:image._id})
+        if(product.image.length == 0){
+            let image = await Image.findOneAndDelete({filename:product.image});
+            console.log(image)
+            await ImageChunk.deleteMany({files_id:image._id})
+        }
         await Product.deleteOne({_id});
         await User.updateMany({}, { $pullAll: { favoriteProducts: [_id] } });
         res.status(200).send({success:true, message:"Product has been deleted successfully"})
